ETH blocks  - scene helper basic structure
